Nach Update auf GCC 10.3
-fcommon ergänzt, sonst hat der Linker alles mehrfach genommen.

Kommentar von
"https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/thread/RYVPP45PMWPXYBBDKXO3CI7YGL7CDQG6/":

Default to -fno-common
A common mistake in C is omitting extern when declaring a global
variable in a
header file. If the header is included by several files it results in
multiple
definitions of the same variable. In previous GCC versions this error is
ignored. GCC 10 defaults to -fno-common, which means a linker error will
now be
reported. To fix this, use extern in header files when declaring global
variables, and ensure each global is defined in exactly one C file. As a
workaround, legacy C code can be compiled with -fcommon.